+++ MonoGame/Players/Remote.cs
@@ -23,7 +23,7 @@ public class Remote : Player
 
     protected override void OnDisplay(IRenderable renderable, Texture2D texture = null, Rectangle? destination = null,
         Rectangle? source = null, Color? color = null, float? rotation = null, Vector2? origin = null,
-        SpriteEffects effect = SpriteEffects.None, int? depth = null)
+        SpriteEffects effect = SpriteEffects.None, float? depth = null)
     {
         _networkClient.Enqueue(renderable, texture, destination, source, color, rotation, origin, effect, depth);
     }
